update tables and else

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('users', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('phone')->unique();
+            $table->bigInteger('phone')->unique();
             $table->boolean('Is_Admin')->default(false);        //edit to false and delete email var
             $table->string('password');
             $table->rememberToken();
